pyproject.toml
[project]
name = "trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.5.46",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langsmith>=0.7.23",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "pytest>=8.0.0",
    "pyright>=1.1.408",
]
